+++ src/common/FooterButton.jsx
@@ -1,4 +1,4 @@
-import "./FooterButton.css";
+import "../styles/FooterButton.css";
 
 export function FooterButton({ children }) {
   return <button className="footerButton">{children}</button>;
