+++ src/common/LoginButton.jsx
@@ -1,4 +1,4 @@
-import "./LoginButton.css";
+import "../styles/LoginButton.css";
 
 export function LoginButton() {
   return <button className="login-button">Login</button>;
