+++ src/common/NavButton.jsx
@@ -1,4 +1,4 @@
-import "./NavButton.css";
+import "../styles/NavButton.css";
 
 export function NavButton({ children }) {
   return <button className="nav-button">{children}</button>;
