+++ src/styles/Avatar.css
