+++ src/styles/Header.css
@@ -1,6 +1,6 @@
-.header {
+.headerContainer {
   margin: 0 auto;
-  width: 1080px;
+  max-width: 1080px;
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -9,4 +9,9 @@
   font-size: 20px;
   font-weight: 700;
   border-bottom: 1px solid grey;
+}
+.headerContents {
+  display: flex;
+  align-items: center;
+  gap: 30px;
 }
